requirements.txt
discord.py
datetime
asyncio
time
json
random